refactor: Change inject.js behaviour
This change completely refactors inject.js into multiple module js files

Beneficts:
- More readable code
- Possible performance improvement
